requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.1
django-bootstrap-form==3.4
django-braces==1.15.0
django-cors-headers==3.10.1
django-crispy-forms==1.14.0
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
Pillow==9.0.0
psycopg2==2.9.3
python-decouple==3.6
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
uvicorn==0.17.0.post1
whitenoise==5.3.0
